requirements.txt
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-environ
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
mysqlclient
pytz==2025.2
PyYAML==6.0.2
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.9.0
geoip2==4.8.0
django-filter
django-countries
django-prometheus
python-logging-loki
python-json-logger